refactor: separate api function code & file
