+++ src/components/Todos/TodoForm/TodoForm.jsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from "react";
 
-import { createTodo } from "api";
+import { createTodo } from "api/queries/todo";
 
 import { SubmitButton } from "components/common/SubmitButton/SubmitButton";
 import * as S from "./styles";
