+++ src/components/Todos/TodoList/TodoList.jsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from "react";
 
-import { getTodos } from "api";
+import { getTodos } from "api/queries/todo";
 
 import { TodoContent } from "../TodoContent/TodoContent";
