+++ src/data/data.json
@@ -29,150 +29,6 @@
       "todo": "과제하기5",
       "isCompleted": false,
       "userId": 1
-    },
-    {
-      "id": 4,
-      "todo": "과제하기5",
-      "isCompleted": false,
-      "userId": 1
-    },
-    {
-      "id": 4,
-      "todo": "과제하기5",
-      "isCompleted": false,
-      "userId": 1
-    },
-    {
-      "id": 4,
-      "todo": "과제하기5",
-      "isCompleted": false,
-      "userId": 1
-    },
-    {
-      "id": 4,
-      "todo": "과제하기5",
-      "isCompleted": false,
-      "userId": 1
-    },
-    {
-      "id": 4,
-      "todo": "과제하기5",
-      "isCompleted": false,
-      "userId": 1
-    },
-    {
-      "id": 4,
-      "todo": "과제하기5",
-      "isCompleted": false,
-      "userId": 1
-    },
-    {
-      "id": 4,
-      "todo": "과제하기5",
-      "isCompleted": false,
-      "userId": 1
-    },
-    {
-      "id": 4,
-      "todo": "과제하기5",
-      "isCompleted": false,
-      "userId": 1
-    },
-    {
-      "id": 4,
-      "todo": "과제하기5",
-      "isCompleted": false,
-      "userId": 1
-    },
-    {
-      "id": 4,
-      "todo": "과제하기5",
-      "isCompleted": false,
-      "userId": 1
-    },
-    {
-      "id": 4,
-      "todo": "과제하기5",
-      "isCompleted": false,
-      "userId": 1
-    },
-    {
-      "id": 4,
-      "todo": "과제하기5",
-      "isCompleted": false,
-      "userId": 1
-    },
-    {
-      "id": 4,
-      "todo": "과제하기5",
-      "isCompleted": false,
-      "userId": 1
-    },
-    {
-      "id": 4,
-      "todo": "과제하기5",
-      "isCompleted": false,
-      "userId": 1
-    },
-    {
-      "id": 4,
-      "todo": "과제하기5",
-      "isCompleted": false,
-      "userId": 1
-    },
-    {
-      "id": 4,
-      "todo": "과제하기5",
-      "isCompleted": false,
-      "userId": 1
-    },
-    {
-      "id": 4,
-      "todo": "과제하기5",
-      "isCompleted": false,
-      "userId": 1
-    },
-    {
-      "id": 4,
-      "todo": "과제하기5",
-      "isCompleted": false,
-      "userId": 1
-    },
-    {
-      "id": 4,
-      "todo": "과제하기5",
-      "isCompleted": false,
-      "userId": 1
-    },
-    {
-      "id": 4,
-      "todo": "과제하기5",
-      "isCompleted": false,
-      "userId": 1
-    },
-    {
-      "id": 4,
-      "todo": "과제하기5",
-      "isCompleted": false,
-      "userId": 1
-    },
-    {
-      "id": 4,
-      "todo": "과제하기5",
-      "isCompleted": false,
-      "userId": 1
-    },
-    {
-      "id": 4,
-      "todo": "과제하기5",
-      "isCompleted": false,
-      "userId": 1
-    },
-    {
-      "id": 4,
-      "todo": "과제하기5",
-      "isCompleted": false,
-      "userId": 1
     }
   ]
 }
